v1.4

+++ dllSrvRunner/dllSrvRunner.cpp
@@ -1,44 +1,85 @@
 #include "pch.h"
 #include "dllSrvRunner.h"
 
-#define ThrowToCS(msg) throw msg
-
-#define SUCCESS                                0L
-#define ERROR_UNKNOWN                      99999L
-#define ERROR_NO_MSG                       10001L
-
-/*
-* * Already defined in winerror.h
-**
-#define ERROR_PATH_NOT_FOUND                   3L
-#define ERROR_ACCESS_DENIED                    5L
-#define ERROR_INVALID_NAME                   123L
-#define ERROR_BAD_EXE_FORMAT                 193L
-#define ERROR_SERVICE_ALREADY_RUNNING       1056L
-#define ERROR_SERVICE_DOES_NOT_EXIST        1060L
-#define ERROR_SERVICE_CANNOT_ACCEPT_CTRL    1061L
-#define ERROR_SERVICE_NOT_ACTIVE            1062L
-#define ERROR_SERVICE_MARKED_FOR_DELETE     1072L
-#define ERROR_SERVICE_EXISTS                1073L
-*/
-
 std::unordered_map<DWORD, PCWSTR> dllSrvRunner::expected_err = {
 {ERROR_PATH_NOT_FOUND,            TEXT("找不到服务二进制文件。")},
 {ERROR_ACCESS_DENIED,             TEXT("拒绝访问。")},
 {ERROR_INVALID_NAME,              TEXT("指定的服务名无效。")},
 {ERROR_BAD_EXE_FORMAT,            TEXT("指定的二进制文件无效。")},
 {ERROR_SERVICE_ALREADY_RUNNING,   TEXT("服务的实例已在运行。")},
+{ERROR_SERVICE_DISABLED,          TEXT("服务已被禁用。")},
 {ERROR_SERVICE_DOES_NOT_EXIST,    TEXT("指定的服务不存在。")},
 {ERROR_SERVICE_CANNOT_ACCEPT_CTRL,TEXT("服务无法在此时接受控制信息。服务可能已停止/正在停止/正在启动。")},
 {ERROR_SERVICE_NOT_ACTIVE,        TEXT("服务尚未启动。")},
 {ERROR_SERVICE_MARKED_FOR_DELETE, TEXT("服务已标记为删除。")},
 {ERROR_NO_MSG,                    TEXT("")} };
-
 std::unordered_map<DWORD, PCWSTR>::iterator dllSrvRunner::errIter;
-
-SC_HANDLE dllSrvRunner::hSCManager;
 std::unordered_map<PWSTR, SC_HANDLE> dllSrvRunner::hSrvMap;
 std::unordered_map<PWSTR, SC_HANDLE>::iterator dllSrvRunner::srvIter;
+
+SC_HANDLE dllSrvRunner::hSCManager;
+
+#pragma region wcsapp
+
+template<typename P, typename Q, typename...R>
+inline void wcsapp(P p, Q q, R...rs)
+{
+    static_assert(0, "[wcsapp] Argument type mismatch.");
+}
+
+template<size_t __Size, typename...R>
+inline void wcsapp(WCHAR(&_Destination)[__Size], PCWSTR _Source, R..._res)
+{
+    wcscat_s(_Destination, _Source);
+    wcsapp(_Destination, _res...);
+}
+
+template<size_t __Size, typename...R>
+inline void wcsapp(WCHAR(&_Destination)[__Size], PWSTR _Source, R..._res)
+{
+    wcscat_s(_Destination, _Source);
+    wcsapp(_Destination, _res...);
+}
+
+template<size_t __Size>
+inline void wcsapp(WCHAR(&_Destination)[__Size], PCWSTR _Source)
+{
+    wcscat_s(_Destination, _Source);
+}
+
+template<size_t __Size>
+inline void wcsapp(WCHAR(&_Destination)[__Size], PWSTR _Source)
+{
+    wcscat_s(_Destination, _Source);
+}
+
+template<typename...R>
+inline void wcsapp(PWSTR(&_Destination), rsize_t _SizeInWords, PCWSTR _Source, R..._res)
+{
+    wcscat_s(_Destination, _SizeInWords, _Source);
+    wcsapp(_Destination, _SizeInWords - wcslen(_Source), _res...);
+}
+
+template<typename...R>
+inline void wcsapp(PWSTR(&_Destination), rsize_t _SizeInWords, PWSTR _Source, R..._res)
+{
+    wcscat_s(_Destination, _SizeInWords, _Source);
+    wcsapp(_Destination, _SizeInWords - wcslen(_Source), _res...);
+}
+
+template<>
+inline void wcsapp(PWSTR(&_Destination), rsize_t _SizeInWords, PCWSTR _Source)
+{
+    wcscat_s(_Destination, _SizeInWords, _Source);
+}
+
+template<>
+inline void wcsapp(PWSTR(&_Destination), rsize_t _SizeInWords, PWSTR _Source)
+{
+    wcscat_s(_Destination, _SizeInWords, _Source);
+}
+
+#pragma endregion
 
 inline void AppendErrInf(PWSTR msg, DWORD errCode, PWSTR end = TEXT("\r\n"))
 {
@@ -47,58 +88,14 @@ inline void AppendErrInf(PWSTR msg, DWORD errCode, PWSTR end = TEXT("\r\n"))
     {
         if (dllSrvRunner::errIter->first != ERROR_NO_MSG)
         {
-            wcscat_s(msg, MSG_BUF_SIZE, dllSrvRunner::errIter->second);
-            wcscat_s(msg, MSG_BUF_SIZE, end);
+            wcsapp(msg, MSG_BUF_SIZE, dllSrvRunner::errIter->second, end);
         }
     }
     else
     {
-        wcscat_s(msg, MSG_BUF_SIZE, TEXT("未知错误。"));
-        wcscat_s(msg, MSG_BUF_SIZE, end);
+        wcsapp(msg, MSG_BUF_SIZE, TEXT("未知错误。"), end);
     }
 }
-
-#pragma region wcsapp
-
-template<size_t __Size, typename P, typename...Q>
-inline void wcsapp(WCHAR(&_Destination)[__Size], P _Source, Q..._res)
-{
-    static_assert(0, "[wcsapp] Argument type mismatch.");
-}
-
-template<size_t __Size, typename...Q>
-inline void wcsapp(WCHAR(&_Destination)[__Size], PWCHAR _Source, Q..._res)
-{
-    wcscat_s(_Destination, _Source);
-    wcsapp(_Destination, _res...);
-}
-
-template<size_t __Size, typename...Q>
-inline void wcsapp(WCHAR(&_Destination)[__Size], PCWCHAR _Source, Q..._res)
-{
-    wcscat_s(_Destination, _Source);
-    wcsapp(_Destination, _res...);
-}
-
-template<size_t __Size, typename P>
-inline void wcsapp(WCHAR(&_Destination)[__Size], P _Source)
-{
-    static_assert(0, "[wcsapp] Argument type mismatch.");
-}
-
-template<size_t __Size>
-inline void wcsapp(WCHAR(&_Destination)[__Size], PWCHAR _Source)
-{
-    wcscat_s(_Destination, _Source);
-}
-
-template<size_t __Size>
-inline void wcsapp(WCHAR(&_Destination)[__Size], PCWCHAR _Source)
-{
-    wcscat_s(_Destination, _Source);
-}
-
-#pragma endregion
 
 STATUS::STATUS(DWORD exitCode, PCWSTR msg)
 {
@@ -182,12 +179,18 @@ STATUS dllSrvRunner::dllLookup(PWSTR srvName)
 
     DWORD cbBufSize = cbBytesNeeded;
     LPQUERY_SERVICE_CONFIG pSrvConf = (LPQUERY_SERVICE_CONFIG)malloc(cbBufSize);
-    if (pSrvConf == NULL || !QueryServiceConfig(hService, pSrvConf, cbBufSize, &cbBytesNeeded))
+    if (pSrvConf == NULL)
     {
+        return STATUS(ERROR_UNKNOWN, TEXT("获取信息时内存申请失败，"));
+    }
+
+    if (!QueryServiceConfig(hService, pSrvConf, cbBufSize, &cbBytesNeeded))
+    {
+        free(pSrvConf);
         return STATUS(GetLastError(), TEXT("获取信息失败，"));
     }
 
-    LPSERVICE_DESCRIPTION pSrvDesc;
+    WCHAR errMsg[64];
     if (
         !QueryServiceConfig2(
             hService,
@@ -199,15 +202,22 @@ STATUS dllSrvRunner::dllLookup(PWSTR srvName)
         errCode = GetLastError();
         if (errCode != ERROR_INSUFFICIENT_BUFFER)
         {
-            goto SET_FAILED_MSG;
+            wcscpy_s(errMsg, TEXT("获取描述失败，"));
+            goto SET_DESC_FAILED_MSG;
         }
     }
 
     cbBufSize = cbBytesNeeded;
-    pSrvDesc = (LPSERVICE_DESCRIPTION)malloc(cbBufSize);
+    LPSERVICE_DESCRIPTION pSrvDesc = (LPSERVICE_DESCRIPTION)malloc(cbBufSize);
+    if (pSrvDesc == NULL)
+    {
+        errCode = ERROR_UNKNOWN;
+        wcscpy_s(errMsg, TEXT("获取描述信息时内存申请失败，"));
+        goto SET_DESC_FAILED_MSG;
+    }
 
-    if (pSrvDesc == NULL
-        || !QueryServiceConfig2(
+    if (
+        !QueryServiceConfig2(
             hService,
             SERVICE_CONFIG_DESCRIPTION,
             (LPBYTE)pSrvDesc,
@@ -216,45 +226,202 @@ STATUS dllSrvRunner::dllLookup(PWSTR srvName)
     {
         errCode = GetLastError();
         free(pSrvDesc);
-        goto SET_FAILED_MSG;
+        wcscpy_s(errMsg, TEXT("获取描述失败，"));
+        goto SET_DESC_FAILED_MSG;
     }
 
     if (0)
     {
-    SET_FAILED_MSG:
+    SET_DESC_FAILED_MSG:
         pSrvDesc = (LPSERVICE_DESCRIPTION)malloc(64 * sizeof(WCHAR));
-        if (pSrvDesc == NULL)
+        pSrvDesc->lpDescription = (LPWSTR)(&pSrvDesc->lpDescription + 1);
+        wcscpy_s(pSrvDesc->lpDescription, 63, errMsg);
+        AppendErrInf(pSrvDesc->lpDescription, errCode, TEXT(""));
+    }
+
+    if (
+        !QueryServiceStatusEx(
+            hService,
+            SC_STATUS_PROCESS_INFO,
+            NULL,
+            0,
+            &cbBytesNeeded))
+    {
+        errCode = GetLastError();
+        if (errCode != ERROR_INSUFFICIENT_BUFFER)
         {
-            return STATUS(GetLastError(), TEXT("内存分配失败，"));
+            wcscpy_s(errMsg, TEXT("获取状态失败，"));
+            goto SET_STATUS_FAILED_MSG;
         }
-        pSrvDesc->lpDescription = (LPWSTR)((LPBYTE)&pSrvDesc->lpDescription + sizeof(LPWSTR));
-        wcscpy_s(pSrvDesc->lpDescription, 63, L"获取描述失败，");
-        AppendErrInf(pSrvDesc->lpDescription, errCode, L"");
+    }
+
+    cbBufSize = cbBytesNeeded;
+    LPSERVICE_STATUS_PROCESS pSrvStatus = (LPSERVICE_STATUS_PROCESS)malloc(cbBufSize);
+    if (pSrvStatus == NULL)
+    {
+        errCode = ERROR_UNKNOWN;
+        wcscpy_s(errMsg, TEXT("获取状态信息时内存申请失败，"));
+        goto SET_STATUS_FAILED_MSG;
+    }
+
+    if (
+        !QueryServiceStatusEx(
+            hService,
+            SC_STATUS_PROCESS_INFO,
+            (LPBYTE)pSrvStatus,
+            cbBufSize,
+            &cbBytesNeeded))
+    {
+        errCode = GetLastError();
+        free(pSrvStatus);
+        wcscpy_s(errMsg, TEXT("获取状态失败，"));
+        goto SET_DESC_FAILED_MSG;
+    }
+
+    if (0)
+    {
+    SET_STATUS_FAILED_MSG:
+        pSrvStatus = (LPSERVICE_STATUS_PROCESS)malloc(sizeof(SERVICE_STATUS_PROCESS));
+        pSrvStatus->dwCurrentState = 0;
+        pSrvStatus->dwProcessId = 0;
+        AppendErrInf(errMsg, errCode, TEXT(""));
     }
 
     WCHAR srvConf[MSG_BUF_SIZE];
+    LPSTR lpServiceType, lpStartType, lpErrorControl;
+    LPWSTR lpCurrentState;
+
+    switch (pSrvConf->dwServiceType)
+    {
+    case SERVICE_FILE_SYSTEM_DRIVER:
+        lpServiceType = "SERVICE_FILE_SYSTEM_DRIVER";
+        break;
+    case SERVICE_KERNEL_DRIVER:
+        lpServiceType = "SERVICE_KERNEL_DRIVER";
+        break;
+    case SERVICE_WIN32_OWN_PROCESS:
+        lpServiceType = "SERVICE_WIN32_OWN_PROCESS";
+        break;
+    case SERVICE_WIN32_SHARE_PROCESS:
+        lpServiceType = "SERVICE_WIN32_SHARE_PROCESS";
+        break;
+    case SERVICE_INTERACTIVE_PROCESS:
+        lpServiceType = "SERVICE_INTERACTIVE_PROCESS";
+        break;
+    default:
+        lpServiceType = "UNKNOWN";
+        break;
+    }
+    switch (pSrvConf->dwStartType)
+    {
+    case SERVICE_AUTO_START:
+        lpStartType = "SERVICE_AUTO_START";
+        break;
+    case SERVICE_BOOT_START:
+        lpStartType = "SERVICE_BOOT_START";
+        break;
+    case SERVICE_DEMAND_START:
+        lpStartType = "SERVICE_DEMAND_START";
+        break;
+    case SERVICE_DISABLED:
+        lpStartType = "SERVICE_DISABLED";
+        break;
+    case SERVICE_SYSTEM_START:
+        lpStartType = "SERVICE_SYSTEM_START";
+        break;
+    default:
+        lpStartType = "";
+        break;
+    }
+    switch (pSrvConf->dwErrorControl)
+    {
+    case SERVICE_ERROR_CRITICAL:
+        lpErrorControl = "SERVICE_ERROR_CRITICAL";
+        break;
+    case SERVICE_ERROR_IGNORE:
+        lpErrorControl = "SERVICE_ERROR_IGNORE";
+        break;
+    case SERVICE_ERROR_NORMAL:
+        lpErrorControl = "SERVICE_ERROR_NORMAL";
+        break;
+    case SERVICE_ERROR_SEVERE:
+        lpErrorControl = "SERVICE_ERROR_SEVERE";
+        break;
+    default:
+        lpErrorControl = "";
+        break;
+    }
+    switch (pSrvStatus->dwCurrentState)
+    {
+    case SERVICE_CONTINUE_PENDING:
+        lpCurrentState = TEXT("正在恢复");
+        break;
+    case SERVICE_PAUSE_PENDING:
+        lpCurrentState = TEXT("正在暂停");
+        break;
+    case SERVICE_PAUSED:
+        lpCurrentState = TEXT("已暂停");
+        break;
+    case SERVICE_RUNNING:
+        lpCurrentState = TEXT("正在运行");
+        break;
+    case SERVICE_START_PENDING:
+        lpCurrentState = TEXT("正在启动");
+        break;
+    case SERVICE_STOP_PENDING:
+        lpCurrentState = TEXT("正在停止");
+        break;
+    case SERVICE_STOPPED:
+        lpCurrentState = TEXT("已停止");
+        break;
+    default:
+        lpCurrentState = errMsg;
+        break;
+    }
+
+    for (PWCHAR ptr = pSrvConf->lpDependencies; ptr < pSrvConf->lpServiceStartName; ptr++)
+    {
+        if (*ptr == L'\0')
+        {
+            if (*(ptr + 1) != L'\0')
+            {
+                *ptr = L',';
+            }
+            else
+            {
+                break;
+            }
+        }
+    }
+
     swprintf_s(srvConf,
         L"服务名称：%ls\r\n"
         L"显示名称：%ls\r\n"
-        L"类型：0x%x\r\n"
+        L"类型：%hs\r\n"
         L"描述：%ls\r\n"
         L"可执行文件的路径：%ls\r\n"
-        L"启动类型：0x%x\r\n"
-        L"错误控制：0x%x\r\n"
+        L"启动类型：%hs\r\n"
+        L"状态：%ls\r\n"
+        L"PID：%d%hs\r\n"
+        L"错误控制：%hs\r\n"
         L"登录身份：%ls\r\n"
         L"分组：%ls\r\n"
-        L"组内标识：%d\r\n"
+        L"组内标识：%d%hs\r\n"
         L"依赖项：%ls\r\n",
         srvName,
         pSrvConf->lpDisplayName,
-        pSrvConf->dwServiceType,
+        lpServiceType,
         pSrvDesc->lpDescription == NULL ? L"" : pSrvDesc->lpDescription,
         pSrvConf->lpBinaryPathName,
-        pSrvConf->dwStartType,
-        pSrvConf->dwErrorControl,
+        lpStartType,
+        lpCurrentState,
+        pSrvStatus->dwProcessId,
+        pSrvStatus->dwProcessId == 0 ? " (N/A)" : "",
+        lpErrorControl,
         pSrvConf->lpServiceStartName,
         pSrvConf->lpLoadOrderGroup == NULL ? L"" : pSrvConf->lpLoadOrderGroup,
         pSrvConf->dwTagId,
+        pSrvConf->dwTagId == 0 ? " (N/A)" : "",
         pSrvConf->lpDependencies == NULL ? L"" : pSrvConf->lpDependencies);
 
     return STATUS(SUCCESS, srvConf);
@@ -337,7 +504,7 @@ STATUS dllSrvRunner::dllStop(PWSTR srvName)
     return STATUS(SUCCESS);
 }
 
-STATUS dllSrvRunner::dllDelete(PWSTR srvName)
+STATUS dllSrvRunner::dllDelete(PWSTR srvName, BOOL force)
 {
     SC_HANDLE hService;
     STATUS ret = dllOpenSrv(hService, srvName);
@@ -345,6 +512,12 @@ STATUS dllSrvRunner::dllDelete(PWSTR srvName)
     {
         return ret;
     }
+
+    if (!force && hSrvMap.find(srvName) == hSrvMap.end())
+    {
+        return STATUS(ERROR_NO_MSG, TEXT("删除操作需要确认。"));
+    }
+
     if (!DeleteService(hService))
     {
         dllDelHandle(hService);
